backend/requirements.txt
asgiref==3.11.1
channels==4.2.2
Django==5.2.12
psycopg2==2.9.11
sqlparse==0.5.5
tzdata==2025.3
python-dotenv==1.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
drf-spectacular==0.28.0
boto3==1.38.46
django-allauth[headless,socialaccount]==65.15.1
Django-Verify-Email==3.0.3
daphne==4.2.1